Switch to simpler onAnyUiInteraction approach
