+++ mathesar_ui/src/components/file-attachments/cell-content/AttachedFileReference.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import { onDestroy, onMount, tick } from 'svelte';
   import { _ } from 'svelte-i18n';
 
   import type { FileManifest } from '@mathesar/api/rpc/records';
@@ -13,11 +12,7 @@
   import Tooltip from '@mathesar/component-library/tooltip/Tooltip.svelte';
   import ContentLoading from '@mathesar/components/ContentLoading.svelte';
   import { toast } from '@mathesar/stores/toast';
-  import {
-    hadInteractionSince,
-    snapshotInteractions,
-    subscribeUIInteractions,
-  } from '@mathesar/utils/subscribeUiInteractions';
+  import { onAnyUiInteraction } from '@mathesar/utils/onAnyUiInteraction';
 
   import { fetchImage, getFileName, getFileViewerType } from '../fileUtils';
 
@@ -41,18 +36,6 @@
   let imageElement: HTMLImageElement | undefined;
   let defaultFileTrigger: HTMLElement;
 
-  // Subscribe to UI interactions to cancel the lightbox opening if the
-  // user interacts with anything _except_ a lightbox trigger.
-  // This also efectively  allows the user to "preload" images.
-  let unsubscribeInteractions: (() => void) | undefined;
-  onMount(() => {
-    unsubscribeInteractions = subscribeUIInteractions([
-      'pointerdown',
-      'keydown',
-    ]);
-  });
-  onDestroy(() => unsubscribeInteractions?.());
-
   $: ({ uri, thumbnail, direct, mimetype } = manifest);
   $: fileViewerType = getFileViewerType(manifest);
   $: thumbnailUrl = `${thumbnail}?height=${thumbnailResolutionHeightPx}`;
@@ -63,6 +46,7 @@
 
   async function loadImage() {
     imageLoading = true;
+    await new Promise((f) => setTimeout(f, 3000));
     imageElement = await fetchImage(direct);
     imageLoading = false;
   }
@@ -72,17 +56,17 @@
     // inactive cell should _not_ open the file viewer.
     if (!canOpenViewer) return;
 
-    // Capture a snapshot of interactions after the image has been clicked,
-    // so we can block opening the lightbox if the user interacts elsewhere.
-    // The image will still be loaded.
-    const snapshot = snapshotInteractions();
-
     if (!imageElement) {
-      await tick();
+      let uiInteraction = false;
+      onAnyUiInteraction(() => {
+        uiInteraction = true;
+      }, ['pointerdown', 'keydown']);
+
       await loadImage();
 
-      // Stop without loading the lightbox if the user has interacted elsewhere.
-      if (hadInteractionSince(snapshot)) return;
+      // If the user has interacted before the lightbox
+      // opened, do not open the lightbox.
+      if (uiInteraction) return;
     }
 
     if (!imageElement) {
